Update 3-Q2.py
Small logic error in starts_with. Writing has_sublist. It's mutually recursive between aux and coaux functions. Aux assumes that sublist IS a sublist, that we're matching part of the sublist with part of the list. Coaux assumes we haven't yet found the first character of the sublist in the list. The posis argument just carries the whole sublist for the period of execution of aux, just in case this particular match is only partial.