Worked on Subsets
It’s closer. Clearly append isn’t the right idea for our combiner
method, but it’s a nice fill in until I think of the right way to go
about this. Basically I want to append them all to the same list. The
obvious way to do this is to accumulate across a passed argument, so I
think that’s what I’ll do next. After that, it should just work
straightforwardly. Also there’s an index error on the first call, since
if n==1 then m==0 so we get nothing, rather than the one item lst[j].